feat(): use separate files for each function in layout templates
